docs: proposed change for module docs
This also removes some more opinionated references. Nix and the Nix daemon can be
configured in a lot of different ways, with configuration not neccessarily aligning
with the expectations previously made, like having a bound name for inputs.

Maybe it is a good idea to explain the difference between `extra-` nix config options,
or having a stateful vs stateless nix.conf, or `trusted-users`, but that seemed more
like a good opportunity for a blog post than a flake readme.

idk